+++ tests/integration/test_authentication.py
@@ -6,7 +6,8 @@ import logging
 from pathlib import Path
 
 import jubilant
-from helpers import connect_cql, get_address, get_secrets_by_label
+from helpers.cassandra import connect_cql
+from helpers.juju import get_address, get_secrets_by_label
 from pytest import raises
 
 logger = logging.getLogger(__name__)
