+++ tests/integration/test_tls.py
@@ -6,8 +6,9 @@ import logging
 from pathlib import Path
 
 import jubilant
-from help_types import IntegrationTestsCharms
-from helpers import check_node_is_up, check_tls, get_address, get_secrets_by_label
+from helpers.cassandra import check_tls
+from helpers.juju import check_node_is_up, get_address, get_secrets_by_label
+from helpers.types import IntegrationTestsCharms
 
 logger = logging.getLogger(__name__)
